requirements.txt
transformers
Pillow
torch
requests
argparse
